Print template parameter lists in print_scope
