+++ src/frontend/cxx-scope.c
@@ -6355,117 +6355,6 @@ int get_template_nesting_of_template_parameters(template_parameter_list_t* templ
     return nesting;
 }
 
-// Debugging
-void print_template_parameter_list_aux(template_parameter_list_t* template_parameters, int* n)
-{
-    if (template_parameters == NULL)
-        return;
-
-    print_template_parameter_list_aux(template_parameters->enclosing, n);
-
-    (*n)++;
-
-    if (template_parameters->num_parameters > 0)
-    {
-        int i;
-        for (i = 0; i < template_parameters->num_parameters; i++)
-        {
-            const char* kind_name = "<<unknown>>";
-            if (template_parameters->parameters[i] != NULL)
-            {
-                switch (template_parameters->parameters[i]->kind)
-                {
-                    case TPK_NONTYPE: kind_name = "nontype"; break;
-                    case TPK_TYPE: kind_name = "type"; break;
-                    case TPK_TEMPLATE: kind_name = "template"; break;
-
-                    case TPK_NONTYPE_PACK: kind_name = "nontype pack"; break;
-                    case TPK_TYPE_PACK: kind_name = "type pack"; break;
-                    case TPK_TEMPLATE_PACK: kind_name = "template pack"; break;
-                    default: break;
-                }
-                fprintf(stderr, "* Nesting: %d | Position: %d | Name: %s | Kind : %s\n", *n, i, 
-                        template_parameters->parameters[i]->entry != NULL
-                        ?  template_parameters->parameters[i]->entry->symbol_name
-                        : "<<unknown symbol>>",
-                        kind_name);
-            }
-            else
-            {
-                fprintf(stderr, "* Nesting: %d | Position: %d | <<unknown parameter>\n", *n, i);
-            }
-
-            template_parameter_value_t* v = template_parameters->arguments[i];
-            if (v == NULL)
-            {
-                fprintf(stderr, "  Argument: <<NONE>>\n");
-            }
-            else
-            {
-                switch (v->kind)
-                {
-                    case TPK_TYPE:
-                    case TPK_TEMPLATE:
-                        {
-                            fprintf(stderr, "  Argument: %s\n", print_declarator(v->type));
-                            break;
-                        }
-                    case TPK_NONTYPE:
-                        {
-                            fprintf(stderr, "  Argument: ");
-
-                            if (nodecl_is_list(v->value))
-                            {
-                                int num_items;
-                                nodecl_t* list = nodecl_unpack_list(v->value, &num_items);
-                                int j;
-                                for (j = 0; j < num_items; j++)
-                                {
-                                    if (j > 0)
-                                        fprintf(stderr, ", ");
-
-                                    fprintf(stderr, "%s", codegen_to_str(list[j],
-                                                CURRENT_COMPILED_FILE->global_decl_context));
-                                }
-
-                                fprintf(stderr, "\n");
-
-                                DELETE(list);
-                            }
-                            else
-                            {
-                                fprintf(stderr, "%s\n", codegen_to_str(v->value,
-                                            CURRENT_COMPILED_FILE->global_decl_context));
-                            }
-
-                            fprintf(stderr, "  (Type: %s)\n", print_declarator(v->type));
-                            break;
-                        }
-                    default:
-                        {
-                            fprintf(stderr, "  Argument: ????\n");
-                            break;
-                        }
-                }
-            }
-        }
-    }
-    else
-    {
-        fprintf(stderr, "* Nesting: %d <<<EMPTY>>>\n", *n);
-    }
-}
-
-void print_template_parameter_list(template_parameter_list_t* template_parameters)
-{
-    if (template_parameters == NULL)
-    {
-        fprintf(stderr, "<<<No template parameters>>>\n");
-    }
-    int n = 0;
-    print_template_parameter_list_aux(template_parameters, &n);
-}
-
 scope_entry_list_t* query_nodecl_name_flags(const decl_context_t* decl_context,
         nodecl_t nodecl_name, field_path_t* field_path, decl_flags_t decl_flags);
